# docs/swagger.yaml
definitions:
  enum.LogStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - Unknown
    - Queued
    - Pending
    - Succeed
    - Failed
  manager.LogMock:
    properties:
      ID:
        type: integer
      console:
        items:
          type: string
        type: array
      hasInfo:
        type: boolean
      processedEpisodes:
        type: integer
      processedFiles:
        type: integer
      status:
        $ref: '#/definitions/enum.LogStatus'
      timeEstimated:
        type: integer
      totalEpisodes:
        type: integer
      totalFiles:
        type: integer
      webhookSend:
        type: boolean
    type: object
  manager.PageScrapRequest:
    properties:
      authorization:
        type: string
      pages:
        items:
          type: string
        type: array
      webhookUrl:
        type: string
    type: object
  manager.PerPageResponse:
    properties:
      logId:
        type: integer
    type: object
  manager.SendWebhookRequest:
    properties:
      authorization:
        type: string
      logIds:
        items:
          type: integer
        type: array
      webhookUrl:
        type: string
    type: object
info:
  contact: {}
  title: Comic Scrapper
  version: "1.0"
paths:
  /api/v1/logger/{logID}:
    get:
      consumes:
      - application/json
      description: Get Request Result By LogID
      parameters:
      - description: logID
        in: path
        name: logID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/manager.LogMock'
      summary: Get Request Result
      tags:
      - Log
  /api/v1/scrapper/request:
    post:
      consumes:
      - application/json
      description: Request Scrapper to begin the process
      parameters:
      - description: Fields
        in: body
        name: fields
        required: true
        schema:
          $ref: '#/definitions/manager.PageScrapRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/manager.PerPageResponse'
            type: array
      summary: Request Scrapper
      tags:
      - Service
  /api/v1/scrapper/sendWebhook:
    post:
      consumes:
      - application/json
      description: Send Webhook by LogID
      parameters:
      - description: Fields
        in: body
        name: fields
        required: true
        schema:
          $ref: '#/definitions/manager.SendWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/manager.PerPageResponse'
            type: array
      summary: Send Webhook
      tags:
      - Service
securityDefinitions:
  BearerAuth:
    description: Bearer [...]
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"